first commit in a long time.  Working version

+++ run_space.py
@@ -1,6 +1,9 @@
 #!/bin/env python2.7
 
 import os
+import argparse
+import itertools
+
 import numpy as np
 
 preamble = '''#!/bin/bash
@@ -36,17 +39,46 @@ def get_suffix(**kwargs):
     
     return suffix
     
+def qsub_case(f, M2, N2):
+    suffix = get_suffix(f=f, M2=M2, N2=N2)
+    rootdir = './simulations/shelfstrat%s' % suffix
     
-for M2 in np.logspace(-6, -8, 3):
-    for N2 in np.logspace(-4, -6, 3):
-        suffix = get_suffix(M2=M2, N2=N2)
-        rootdir = './simulations/shelfstrat%s' % suffix
-        f = open('shelfstrat%s.sh' % suffix, 'w')
-        f.writelines(preamble)
-        f.write('python2.7 ./run_case.py --rootdir %s --M2 %4.2e --N2 %4.2e\n' % (rootdir, M2, N2))
-        f.write('/usr/mpi/gcc/openmpi-1.4.3/bin/mpiexec -np 8 ./project/coawstM %s/ocean_shelfstrat.in > %s/ocean_shelfstrat.out\n' % (rootdir, rootdir))
-        f.close()
+    Ri = N2 * f**2 / M2**2
+    phi = M2 * 1e-3 * np.sqrt(Ri) / f**2
+    
+    print 'M2 = %4.2e;   N2 = %4.2e;    f = %4.2e;    Ri=%4.2f;   phi=%4.2f' % (M2, N2, f, Ri, phi)
+    if not args.dryrun:
+        qfile = open('shelfstrat%s.sh' % suffix, 'w')
+        qfile.writelines(preamble)
+        qfile.write('python2.7 ./run_case.py --rootdir %s --M2 %4.2e --N2 %4.2e --f %f \n' % (rootdir, M2, N2, f))
+        qfile.write('/usr/mpi/gcc/openmpi-1.4.3/bin/mpiexec -np 8 ./project/coawstM %s/ocean_shelfstrat.in > %s/ocean_shelfstrat.out\n' % (rootdir, rootdir))
+        qfile.close()
         os.system('qsub shelfstrat%s.sh' % suffix)
 
+
+
+parser = argparse.ArgumentParser()
+parser.add_argument('--M2', type=float, nargs='+',
+                   help='Horizontal stratification')
+parser.add_argument('--Ri', type=float, nargs='+',
+                   help='Richardson number')
+parser.add_argument('--N2', type=float, nargs='+',
+                   help='Vertical stratification')
+parser.add_argument('--f', type=float, nargs='+',
+                   help='Coreolis parameter')
+parser.add_argument('--dryrun', action="store_true")
+args = parser.parse_args()
+
+if args.Ri is None:
+    for M2, N2, f in itertools.product(args.M2, args.N2, args.f):
+        qsub_case(f, M2, N2)
+elif args.M2 is None:
+    for Ri, N2, f in itertools.product(args.Ri, args.N2, args.f):
+        M2 = np.sqrt(N2 * f**2 / Ri)
+        qsub_case(f, M2, N2)
+elif args.N2 is None:
+    for Ri, M2, f in itertools.product(args.Ri, args.M2, args.f):
+        N2 = Ri * M2**2 / f**2
+        qsub_case(f, M2, N2)
